src/app/requirements.txt
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
python-dotenv>=0.19.0
python-magic>=0.4.24
langchain-google-genai>=0.1.0
google-generativeai>=0.3.0
langchain-core>=0.1.0
langchain-community>=0.0.10
python-multipart>=0.0.5
pinecone>=7.3.0,<8.0.0